Adding movement and position based structs
position structs for leg and body IK are added prior to coming up with the
IK scheme or each leg w.r.t to the body of the robot.

Currently using a 32bit fixed point value for position values giving us about 4 digits of
accuracy. This can be shifted up later if we'd like to have the mantissa to be larger.

Need to translate position to servo angles given the gait and constraints applied to that.