Added naive implementation adapted from hexapod IK to Quad config.
Grabbed some code that did stuff for a hexapod, but adapted it to be the quad
by modifying the initial foot position angles as well as the angles related
to the coxa positions.

This can be improved with some precalculation of variables to save us a bunch
of floating point operations.

Curious if there's also the matrix approach that's similar to this / more
efficient. Need to do more digging. Ive got a matrix from Foot to Coxa, then
Body to Coxa but not sure if that's the best way to go about solving this
without Ill conditioning.